examples/python_scripts/requirements.txt
paho-mqtt==1.2.1

